Update WordsFallingToTheGround
"...He let none of Samuel’s words fall to the ground."  1 Samuel 3:19
A reminder to all of us that our words, though they contain the power of death and life, might be powerless for good if the Lord is not supportive of the one speaking.
"Without me ye can do nothing" John 15:5 is another witness to this fact.  This statement by Jesus also implies that we have a desire to accomplish - to do something.  Jesus appeals to our desire to accomplish and does not hold against us that we might seek Him as one means to that end.