Update WordsFallingToTheGround
"...He let none of Samuel’s words fall to the ground."  1 Samuel 3:19
A reminder to all of us that our words, though they contain the power of death and life, might be powerless for good if the Lord is not supportive of the one speaking.
"Without me ye can do nothing" John 15:5 is another witness to this fact.  This statement by Jesus also implies that we have a desire to accomplish - to do something.  Jesus appeals to our desire to accomplish and does not hold against us that we might seek Him as one means to that end.  Another thing to notice here is that Jesus is not speaking to unbelievers in this metter of doing something.  There is a difference between what an unbeliever thinks is an accomplishment and what a sold-out believer does.  Jesus speaks to accomplishing the things that a sold-out believer wants to accomplish - you can't accomplish those things without Him.
"You will seek Me and find Me when you search for Me with all your heart" Jer. 29:!3
Look not so much at the price to pay in this verse, but focus on the fact that there is a promise here that God solemnly makes.  He will without fail honor that promise.  Know that.  But be willing to miss a meal if need be: "Therefore also now, saith the LORD, turn ye even to me with all your heart, and with fasting, and with weeping, and with mourning:"  Joel 2:12